ref_gl/gl_mesh.py
"""
gl_mesh.py - MD2 model mesh rendering
Handles vertex interpolation and frame animation for alias models
"""

from OpenGL.GL import *
from wrapper_qpy.decorators import TODO
from wrapper_qpy.custom_classes import Mutable
from wrapper_qpy.linker import LinkEmptyFunctions

LinkEmptyFunctions(globals(), ["Com_Printf"])


# ===== MD2 Rendering =====

def R_DrawAliasModel(ent):
    """Draw MD2 (alias) model"""
    try:
        if not ent or not ent.model:
            return

        if ent.model.type != 2:  # MODEL_ALIAS
            return

        # TODO: Draw MD2 model with proper animation

    except Exception as e:
        print(f"R_DrawAliasModel error: {e}")


def GL_DrawAliasFrameLerp(paliashdr, backlerp):
    """Draw alias model with frame interpolation"""
    try:
        # Linear interpolation between frames
        # TODO: Get current and next frame
        # TODO: Interpolate vertex positions
        # TODO: Draw mesh

        glBegin(GL_TRIANGLES)
        # TODO: Submit vertices
        glEnd()

    except Exception as e:
        print(f"GL_DrawAliasFrameLerp error: {e}")


def GL_LerpVerts(nverts, v, ov, verts, lerp, move, frontv, backv):
    """Interpolate vertex positions between frames"""
    try:
        # Linear interpolation: v = ov * (1 - lerp) + v * lerp
        for i in range(nverts):
            # TODO: Interpolate each vertex
            pass

    except Exception as e:
        print(f"GL_LerpVerts error: {e}")


def GL_DrawAliasShadow(paliashdr, posenum):
    """Draw shadow beneath alias model"""
    try:
        # TODO: Draw shadow polygon

        glBegin(GL_POLYGON)
        # TODO: Submit shadow vertices
        glEnd()

    except Exception as e:
        print(f"GL_DrawAliasShadow error: {e}")


def GL_LightVertex(v, lv):
    """Apply lighting to vertex"""
    # TODO: Look up lightstyle and apply
    pass


def GL_DrawAliasFrame(paliashdr, posenum):
    """Draw single alias frame without interpolation"""
    try:
        # TODO: Draw frame directly

        glBegin(GL_TRIANGLES)
        # TODO: Submit frame vertices
        glEnd()

    except Exception as e:
        print(f"GL_DrawAliasFrame error: {e}")


@TODO
def GL_DrawAliasMeshFrame(frame):
    pass


@TODO
def GL_DrawAliasModel_MD5(ent):
    """Draw MD5 skeletal model (future)"""
    pass


@TODO
def Mod_LoadAlias(mod):
    pass


@TODO
def Mod_LoadMD2(mod):
    pass


@TODO
def Mod_LoadMD5(mod):
    pass


@TODO
def GL_LerpVerts(nverts, v, ov, verts, lerp, move, frontv, backv):
    pass


@TODO
def GL_DrawAliasFrameLerp(paliashdr, backlerp):
    pass


@TODO
def GL_DrawAliasShadow(paliashdr, posenum):
    pass


@TODO
def R_CullAliasModel(bbox, e):
    pass


@TODO
def R_DrawAliasModel(e):
    pass
