requirements.txt
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7
Pillow==10.3.0
numpy==1.26.4
